requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
celery[redis]==5.2.7
flower==1.2.0
djoser==2.0.5
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
